fixed a few signature problesm with the interface

+++ src/Memory.h
@@ -6,7 +6,7 @@
 class Memory {
 public:
   uint8_t read8(uint16_t);
-  uint8_t read16(uint16_t);
+  uint16_t read16(uint16_t);
 
   void write8(uint16_t, uint8_t);
   void write16(uint16_t, uint16_t);
